Stahovanie kazdu sekundu

+++ src/main/java/com/example/BtcRate/model/RateEntity.java
@@ -2,7 +2,6 @@ package com.example.BtcRate.model;
 
 
 import lombok.*;
-import org.hibernate.validator.constraints.UniqueElements;
 
 import javax.persistence.*;
 
@@ -14,14 +13,13 @@ import javax.persistence.*;
 @ToString
 public class RateEntity extends AbstractEntity{
     @Column(name = "VALUE", nullable = false)
-    private long last;
-    private long high;
-    private long low;
-    private long amount;
-    private long bid;
-    private long ask;
-    private long change;
-    private long open;
-    private long timestamp;
-
+    private double last;
+    private double high;
+    private double low;
+    private double amount;
+    private double bid;
+    private double ask;
+    private double change;
+    private double open;
+    private double timestamp;
 }
